PgpBlockViewQuoteModule
